+++ CHANGELOG.md
@@ -4,6 +4,15 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/) and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html)
 
+## [UNRELEASED]
+
+### Fixed/Changed
+
+- Optimizations:
+  - Construction of the available locales array on first access.
+  - Indices used for language resolution and language aliases.
+  - Conversion of language tags to "fast tag" representation internally
+
 ## [1.0.2]
 
 ### Fixed/Changed