+++ CHANGELOG.md
@@ -4,7 +4,7 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/) and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html)
 
-## [UNRELEASED]
+## [1.0.3]
 
 ### Added
 
@@ -662,6 +662,7 @@ Split locale-related code into 3 new packages which can be used independently of
 
 - Converted `CLDR` to an interface
 
+[1.0.3]: https://github.com/phensley/cldr-engine/compare/v1.0.2...v1.0.3
 [1.0.2]: https://github.com/phensley/cldr-engine/compare/v1.0.1...v1.0.2
 [1.0.1]: https://github.com/phensley/cldr-engine/compare/v1.0.0...v1.0.1
 [1.0.0]: https://github.com/phensley/cldr-engine/compare/v0.26.0...v1.0.0